+++ code/models/mnist_noise_baseline_cnn.py
@@ -56,7 +56,7 @@ def transform_train(img):
 
 def transform_test(img):
     img = np.array(img)
-    img = random_noise(img, mode='salt', seed=RNG, amount=0.1)
+    img = np.float32(random_noise(img, mode='salt', seed=RNG, amount=0.1))
     return transforms_norm(img)
